# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: ADA ASM API
  version: 0.1.0
  description: |
    REST API for ADA ASM. This file is the source of truth for the API
    contract and is regenerated from the FastAPI app's `/openapi.json`
    on every PR that changes `backend/app/api/**`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/components:
    get:
      tags: [components]
      summary: List components (paginated, searchable)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Case-insensitive search over mpn / sku / name / family.
        - in: query
          name: family
          schema: { type: string }
        - in: query
          name: supplier
          schema: { type: string }
        - in: query
          name: tier
          schema: { type: string, enum: ["A+", A, B, C, D] }
        - in: query
          name: nato_score
          schema:
            type: string
            enum: [100_otan, otan, allied_otan, neutral, high_risk, no_otan]
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Paginated catalogue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedComponents" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/Validation"
    post:
      tags: [components]
      summary: Create a component
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComponentCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Component" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: "MPN already registered (code: MPN_ALREADY_REGISTERED)"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          $ref: "#/components/responses/Validation"
...
  /api/v1/customers:
    get:
      tags: [customers]
      summary: List all customers (no pagination yet)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of Customer rows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      tags: [customers]
      summary: Create a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerCreate" }
      responses:
        "201":
          description: Customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "409":
          description: CUSTOMER_HOLDED_ID_ALREADY_REGISTERED
  /api/v1/customers/{customer_id}:
    parameters:
      - in: path
        name: customer_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [customers]
      summary: Single customer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer
        "404":
          description: CUSTOMER_NOT_FOUND
    patch:
      tags: [customers]
      summary: Partial update
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerUpdate" }
      responses:
        "200":
          description: Customer
        "404":
          description: CUSTOMER_NOT_FOUND
        "409":
          description: CUSTOMER_HOLDED_ID_ALREADY_REGISTERED
    delete:
      tags: [customers]
      summary: Delete a customer (projects.customer_id becomes NULL)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
...
  /api/v1/components/lookup:
    get:
      tags: [components]
      summary: Look up a component by MPN across all enabled suppliers
      description: >
        Walks `SUPPLIER_LOOKUP_PRIORITY` and merges quotes progressively
        (first non-null wins per scalar field). Empty strings from a
        higher-priority supplier are treated as missing so a later
        supplier with a real value can fill the gap. Result is cached in
        Redis under `supplier_lookup:{lower(mpn)}` for
        `SUPPLIER_LOOKUP_CACHE_TTL_SECONDS` (default 900). Use this to
        pre-fill the "Nuevo componente" form by MPN. Introduced by the
        `supplier-sync` change.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: mpn
          required: true
          schema: { type: string, minLength: 3, maxLength: 60 }
          description: Manufacturer part number (case-insensitive, trimmed).
        - in: query
          name: force_refresh
          required: false
          schema: { type: boolean, default: false }
          description: Bypass the Redis cache and re-query every enabled supplier.
      responses:
        "200":
          description: Lookup succeeded — at least one supplier returned data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LookupResponse" }
        "401": { description: Unauthenticated. }
        "404":
          description: At least one supplier responded but none had a match.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  code: { type: string, enum: [COMPONENT_MPN_NOT_FOUND] }
        "422":
          description: MPN failed validation (too short, too long, etc.).
        "502":
          description: Every consulted supplier raised a transport error.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  code: { type: string, enum: [SUPPLIER_LOOKUP_UNAVAILABLE] }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 Problem Details with a stable `code` discriminator.
      required: [title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
    Component:
      type: object
      required: [id, mpn, name, family, stock, tier, nato_score, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        mpn: { type: string, maxLength: 100 }
        sku: { type: string, nullable: true, maxLength: 100 }
        name: { type: string, maxLength: 200 }
        family: { type: string, maxLength: 100 }
        description: { type: string, nullable: true }
        datasheet_url: { type: string, nullable: true }
        location: { type: string, nullable: true, maxLength: 100 }
        fabricante: { type: string, nullable: true, maxLength: 120 }
        tipo_almacenamiento:
          type: string
          nullable: true
          maxLength: 80
          description: FE-enforced enum `Gaveta` | `Almacén` (BE column is free-text).
        holded_id: { type: string, nullable: true, maxLength: 80 }
        fecha_creacion: { type: string, format: date, nullable: true }
        notas: { type: string, nullable: true }
        stock: { type: integer, minimum: 0 }
        stock_min:
          type: integer
          nullable: true
          minimum: 0
          description: Explicit floor; when null the effective minimum is `tier * 5`.
        tier:
          type: integer
          enum: [1, 2, 3, 4]
          description: Criticality, 1 = most critical. Cached from active scoring.
        nato_score:
          type: string
          enum: ["A+", A, B, C, D, F]
          description: Cached from active scoring.
        country_of_origin:
          type: string
          nullable: true
          pattern: "^[A-Z]{2}$"
          description: ISO 3166-1 alpha-2.
        proveedor_preferente_id: { type: string, format: uuid, nullable: true }
        current_price_per_100_eur:
          type: string
          nullable: true
          description: Server-computed (latest 100u price from preferred supplier).
        supplier_stock_summary:
          type: array
          description: |
            Latest snapshot per supplier — server-hydrated in a single batched query.
            Drives the `<StockStatusBadge>` colour and tooltip on list/hierarchy tables
            without an N+1 round trip. Empty array means no supplier inventory has ever
            been recorded (not "unknown").
          items:
            $ref: "#/components/schemas/SupplierStockSummaryEntry"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SupplierStockSummaryEntry:
      type: object
      required: [supplier_id, supplier_name, quantity]
      properties:
        supplier_id: { type: string, format: uuid }
        supplier_name: { type: string }
        quantity: { type: integer, minimum: 0 }
...
    ComponentCreate:
      type: object
      required: [mpn, name, family, tier, nato_score]
      properties:
        mpn: { type: string, minLength: 1, maxLength: 100 }
        name: { type: string, minLength: 1, maxLength: 200 }
        family: { type: string, minLength: 1, maxLength: 100 }
        sku: { type: string, nullable: true, maxLength: 100 }
        description: { type: string, nullable: true }
        datasheet_url: { type: string, nullable: true }
        location: { type: string, nullable: true, maxLength: 100 }
        fabricante: { type: string, nullable: true, maxLength: 120 }
        tipo_almacenamiento: { type: string, nullable: true, maxLength: 80 }
        holded_id: { type: string, nullable: true, maxLength: 80 }
        fecha_creacion: { type: string, format: date, nullable: true }
        notas: { type: string, nullable: true }
        stock: { type: integer, minimum: 0, default: 0 }
        stock_min: { type: integer, minimum: 0, nullable: true }
        tier: { type: integer, enum: [1, 2, 3, 4] }
        nato_score: { type: string, enum: ["A+", A, B, C, D, F] }
        country_of_origin: { type: string, nullable: true, pattern: "^[A-Z]{2}$" }
        proveedor_preferente_id: { type: string, format: uuid, nullable: true }
...
    PaginatedComponents:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Component" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
...
    Customer:
      type: object
      description: Id-link to Holded; the FE renders `holded_id` as a clickable badge.
      required: [id, holded_id, name, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        holded_id: { type: string, maxLength: 64 }
        name: { type: string, maxLength: 200 }
        holded_url: { type: string, nullable: true, maxLength: 500 }
        notas: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CustomerCreate:
      type: object
      required: [holded_id, name]
      properties:
        holded_id: { type: string, minLength: 1, maxLength: 64 }
        name: { type: string, minLength: 1, maxLength: 200 }
        holded_url: { type: string, nullable: true, maxLength: 500 }
        notas: { type: string, nullable: true }
    CustomerUpdate:
      type: object
      properties:
        holded_id: { type: string, minLength: 1, maxLength: 64 }
        name: { type: string, minLength: 1, maxLength: 200 }
        holded_url: { type: string, nullable: true, maxLength: 500 }
        notas: { type: string, nullable: true }
...
    SupplierCode:
      type: string
      enum: [mouser, digikey, tme, farnell, rs]
    LookupFields:
      type: object
      description: >
        Merged "headline" view used to pre-fill the new-component form.
        Each field is the first non-null + non-empty value found in
        priority order across the consulted suppliers.
      properties:
        name: { type: string, nullable: true }
        description: { type: string, nullable: true }
        manufacturer: { type: string, nullable: true }
        family_hint: { type: string, nullable: true }
        datasheet_url: { type: string, nullable: true }
        package: { type: string, nullable: true }
        current_price_per_100_eur:
          type: number
          format: float
          nullable: true
          description: >
            EUR cost for 100 units, computed from the largest price break
            whose `quantity <= 100`. NULL when no supplier returned a
            qualifying break.
    SupplierPriceBreakResponse:
      type: object
      required: [quantity, price_original, currency_original]
      properties:
        quantity: { type: integer, minimum: 1 }
        price_original: { type: number, format: float }
        currency_original: { type: string, minLength: 3, maxLength: 3 }
        price_eur: { type: number, format: float, nullable: true }
    SupplierData:
      type: object
      required: [supplier]
      properties:
        supplier: { $ref: "#/components/schemas/SupplierCode" }
        supplier_sku: { type: string, nullable: true }
        supplier_product_url: { type: string, nullable: true }
        stock: { type: integer, nullable: true }
        price_breaks:
          type: array
          items: { $ref: "#/components/schemas/SupplierPriceBreakResponse" }
    LookupResponse:
      type: object
      required: [mpn, found, fields, supplier_data, sources_consulted, sources_succeeded, missing_fields]
      properties:
        mpn: { type: string }
        found: { type: boolean }
        fields: { $ref: "#/components/schemas/LookupFields" }
        supplier_data:
          type: array
          items: { $ref: "#/components/schemas/SupplierData" }
        sources_consulted:
          type: array
          items: { $ref: "#/components/schemas/SupplierCode" }
        sources_succeeded:
          type: array
          items: { $ref: "#/components/schemas/SupplierCode" }
        missing_fields:
          type: array
          items: { type: string }
...
  responses:
    Unauthorized:
      description: |
        Authentication failed. `code` is one of:
        `INVALID_CREDENTIALS`, `UNAUTHENTICATED`, `ACCESS_TOKEN_EXPIRED`,
        `ACCESS_TOKEN_WRONG_TYPE`, `ACCESS_TOKEN_MALFORMED`,
        `REFRESH_TOKEN_EXPIRED`, `REFRESH_TOKEN_REVOKED`,
        `REFRESH_TOKEN_MALFORMED`.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Validation:
      description: Pydantic validation error.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...